test: move prompt registration coverage to contract layer
